move guest and host apis in same module into separate files for better readability
