+++ crates/lunatic-sqlite-api/src/guest_api/mod.rs
@@ -1,0 +1,1 @@
+pub mod sqlite_guest_bindings;
